Add isFull() method to DrugList. Change id data type to long. Add ps.close() and rs.close() to all repository methods. Add update() method to AdminRepository. Add create() and update() methods to PatientRepository. Update create() method of PrescriptionRepository. Add read() method to DrugRepository. Update setPrescriptionStatus(), setDrugExist() and setDrugPrice() methods of AdminService. Add setPrice(), setDoesExist() and isExist() method to DrugService. Add add() method to PatientService. Add load(), loadAll(), setPrescriptionStatus() and isExist() methods to PrescriptionService. create PatientMenu and AdminMenu to menu package. Add login() to AdminMenu. Add signUp() to PatientMenu. Add addPrescription() to PatientMenu. Update Print class.

+++ src/main/java/org/repository/drug/DrugRepository.java
@@ -6,7 +6,7 @@ import org.entity.DrugList;
 import java.sql.SQLException;
 
 public interface DrugRepository {
-void add(Drug drug, int prescriptionId) throws SQLException;
+void add(Drug drug, long prescriptionId) throws SQLException;
     void remove(int id) throws SQLException;
     void update(Drug drug) throws SQLException;
     DrugList read(long prescriptionId) throws SQLException;
